fix(cli): remove unnecessary u32 casts in fs ls/stat permission formatting (#1518)
PR #1507 introduced redundant `as u32` casts when calling format_permission
on FileStatus.mode, which is already u32. This triggers clippy::unnecessary_cast
with -D warnings and blocks the fast profile build gate.

Fixes #1517

+++ curvine-cli/src/cmds/fs/ls.rs
@@ -320,7 +320,7 @@ async fn print_file_entry(
     config: &PrintConfig,
 ) -> CommonResult<()> {
     let file_type = if file.is_dir { "d" } else { "-" };
-    let permissions = crate::cmds::fs::common::format_permission(file.mode as u32);
+    let permissions = crate::cmds::fs::common::format_permission(file.mode);
     let replicas = if file.is_dir {
         "-"
     } else {
